move color animation to own file/mixin
