+++ lib/views/vote/up_vote_button.dart
@@ -4,6 +4,7 @@ import 'package:pr0gramm/entities/enums/vote_button_type.dart';
 import 'package:pr0gramm/services/vote_animation_service.dart';
 import 'package:pr0gramm/views/vote/vote_button.dart';
 import 'package:pr0gramm/views/vote/vote_button_animation_integration.dart';
+import 'package:pr0gramm/views/vote/vote_button_color_animation.dart';
 import 'package:pr0gramm/views/vote/vote_button_rotate_animation.dart';
 
 class UpVoteButton extends VoteButton {
@@ -27,6 +28,7 @@ class _UpVoteButtonState extends State<UpVoteButton>
     with
         TickerProviderStateMixin,
         VoteButtonAnimationIntegration<UpVoteButton>,
+        VoteButtonColorAnimation<UpVoteButton>,
         VoteButtonRotateAnimation<UpVoteButton> {
   @override
   void initState() {
